packaging user defined code
Add the lesson at: node_modules/lib/lessson/unordered_array_lesson.js;
Provides the logic to do a step-by-step traversal of how to insert a
value into an unordered array.

Added testing for the lesson at
node_modules/lib/test/test_unordered_array_lesson.js; Use `npm test` and
open node_modules/lib/test/test.html to run tests.

Modified unorderedArray.html to insert testing code and establish a
workflow to include user defined code; See README for details.

node_modules/lib/wrapper.js is used to instantiate the lesson and pass
in the (mock for now) UI variable used to manipulate the UI.

node_modules/lib/test/mock_UI.js is present for testing the lesson as it
was built in the absence of a UI to test again, will be phased out in
future iterations.

node_modules/lib/test/index.js is used to require() test files to run
with tape.

Added package.json to allow `npm install` to install dependencies of the
lesson logic, as well as define an `npm test` for testing and `npm
start` to build the user defined code that is included with the Animate
code.

.gitignore is defined to ignore files in node_modules except for the
code in `lib`.