remove all JS code from AA
Created lesson_UI to manipulate the UI elements from AA as exposed
through the `lib` variable. lesson_UI is a state based approach to
manipulating the UI and uses the functions as previously defined in AA
code to do that manipulation. All JS code (except for `this.stop()`) has
been removed from AA and placed into lesson_UI, as well as streamlined
as much as possible. Testing for lesson_UI can be found in
node_modules/lib/test/test_lesson_UI.js and can be ran by `npm test` and
opening node_modules/lib/test/test.html in a browser.

Modifications to unorderedArray.html to call lesson_UI are highlighted
with CUSTOM-CODE comments and are placed as they are given the event
based model of execution of the generated AA code.