added arrow_UI, a class that implements a logical iterator and UI arrow element
(also adding array_UI at this time, but unfinished, next commit should
be the finished array_UI class)

Arrow_UI, a class use to implement a UI element of an arrow (with an
attached label)

Each arrow should implement the iterator protocol
(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Iterators_and_Generators).

Each arrow is associated with a data structure and as such requires that
some of its methods be overrident via options passed to the constructor.

Each arrow is decorated with a _ namespace that contains helpful
properties as well as the move() function, to implement moving the arrow
around the stage.