added push() method to array
push(): tweens a createjs Text object to the first available position in
the array (as indexed by its "tail" arrow), inserts that value into the
container at that index (that houses the ArrayElement object), adds the
numeric value of that Text object to a decorated "values" arrow and
returns the new length of the array.

Notes have been made concerning the odd bugs resulting from createjs's
unique local (to a container) vs global (the stage) positioning and how
it impacts positioning of objects that are added to a container and the
tweening of those objects around the stage a simpacted by the
positioning (as above).

Modified the lesson code to account for the array and arrow being
bundled together (to allow for easier positioning of the arrow and
tweening).