added the ability to test with the live AA code
modified `npm test` to dump the tests to bundle.js in the root
directory, this allows exposure of the actual `lib`, `createjs` and
`stage` objects used in the program to be used by the testing
infrastructure.

Modified `node_modules/lib/test/index.js` to be similar to wrapper.js
(in that it populates a decorated namespace into lib and runs
lesson_UI), but then calls the testing code, while making available the
above objects in primitives.