port lesson code to UI proxy object
added a proxy object from the UI, that the lesson can use to manipulate
the UI

created a method to wipe the text input then create a matching
createjs.Text object on the stage that can tweened to the array.

migrated testing to allow tests to use live AA code if needed + migrated
tests to create a bundle.js that can be loaded by the
unorderedArray.html for testing.

currently porting existing lesson logic to UI proxy use:

added a check for is_moving() to all button presses that suppresses
traversal of the lesson or animations if there is an existing animation
happening.

reduced initial array allocation to 2 down from 5, as when creating a
new array a new array of size 4 would fit on the current UI, while 10
would not.

added a function to array_UI that would return a UI arrow that would act
like an iterator over the array, while tweening itself to the next
position in / after the array on next().

debugged and resolved with createjs positioning and adapted
array._.push() to allow tweening from the stage to an array or between
arrays.

modified the arrows to allow pointing to the position after the end of
the array to allow a full traversal of the array and stopping when the
arrows were pointing at the same "spot" (a position after the end of the
array), rather than attempting to create special logic that would handle
both copying a value from the final position in the array AND signaling
termination of the copying logic.