linear search
Created a build directory to house snapshots of lessons on completion,
as their own self contained units.

Created a linear search lesson.  Allocates an array, filled with random
numbers. Allows the user to input a value to search for and walks
through the process of a linear search of the array, either showing that
the search did not locate the number or that it did and why.