abstracted linear search from its lesson
The lesson logic can be abstracted form the lesson which it runs.

So I've made a "logic" sub-directory under the "lesson" directory that
will contain the logic for the lessons going forward (e.g. a lesson will
be an instantiation of a configured stage and some logic to work through
a lesson).