modified linear search lesson logic
If the linear search algorithm locates the value it's searching for,
then the index of that value is returned at the end of the lesson; Else
if it's not found, it returns undefined for the index.

Linear search now requires a "value" argument to search for and no
longer gathers the value from the user input directly (this is left up
to the lesson manager).

Modified the linear search lesson manager to:
* gather the value from the user input (with pre_lesson)
* disable the user input on lesson start
* re-enable the user input on lesson end
* allocate the lesson logic after fetching the user input, on a next()
* re-enter the READY state after the lesson logic has concluded (ready
to allocate another instance of the logic and do another search)

Modified the pre_lesson logic to no longer disable the user input, this
is left up to the lesson manager.