migrated arrow_UI to array_arrow_UI
Where the UI arrow instantiated by array_arrow_UI is an arrow with
several properties that can be used in cojunction with arrays.

Tested and working at time of writing set() (which sets an arrow to an
arbitrary postion in an array [within bounds]), next() (increments to
the next position in the array [if there is one], returning an iterate
like object that contains the value at that position in the array [if
there is one] and a property `done` that indicates if the iteration
through the array is finished) and prev() (which is used similarly to
next(), except it iterates backwards through the array, return values
and has its own return value property of `done`).

These methods are added to support additional functionality required by
array_delete.