added move_within method to array_UI
the method is used to relocate UI ArrayElements within the array, this
coupled with pop() should provide enough functionality to remove an item
from the array, then move other values backwards through the array to
show the process of shrinking the array on deletion (for lesson array
delete).

(also misc changes to tests:

test_array_UI, adapted to use the live UI for testing rather than mocks
test_move, some issues with tests that rely on timing; Either JS or TAP
is scheduling the tests such that asynchronous events (E.g. tweens) are
being executed outside of the tests parameters, causing issues with
tests that check on the state of the async process while it's processing
(e.g. tweening form 0 to 1, half way through the time it takes to tween
the value should be between 0 and 1) and showing the async processes
have completed during the testing process, when they should not (and
have not previously, with fewer tests and no other changes).