_cleanup method in array_UI
Added to manage null values in array._.values array, due to changes in
the array from pop() and move_within() (as both put in `null` for
indices in the array in which they have either removed an item [pop()]
or relocated an item [move_within()]).

Minor bug fixes to tests, as there are so many tests that some async
tests are interfering with each other in a non-deterministic way;
Limitation of the testing environment.