complete unordered array delete lesson
Placed in "unordered array - delete" in the BUILD directory.

The lesson takes a single number, executes a linear search, then if
found, removes the value at the found index, decrements the position of
all values in the array above the removed index, then once reaching the
index that the 'tail' arrow is pointing at, stops copying and decrements
the 'tail' arrow, then finishes the lesson.