binary search lesson logic
An implementation of the binary search algorithm in lesson logic, going
forward we'll use this logic as part of several lessons: insertion into
an ordered array, deletion from an ordered array and just searching in
an ordered array.

The algorithm allows the "low" and "high" arrow to traverse off of the
array, to show how the algorithm behaves in "not found" cases.

The lesson logic also returns the insertion index of where a value could
be added to the array to maintain ordering, on a "not found" case, as
well as reporting to the student.