ordered array insert
Created a lesson to go through the algorithm of:

Doing a binary search to find where to insert in the array
then
Moving values in the array to make room to insert the value to keep the
array ordered

The lesson detects when it's array is full and terminates.

Add the lesson to the BUILD directory.